Refactoring app.py and breaking functionalitiy
